# Adds simd vector extensions functionality to optimize shasumming