… vendor/github.com/minio/sha256-simd/LICENSE
… vendor/github.com/minio/sha256-simd/README.md
… vendor/github.com/minio/sha256-simd/cpuid.go
… vendor/github.com/minio/sha256-simd/cpuid_386.go
… vendor/github.com/minio/sha256-simd/cpuid_386.s
… vendor/github.com/minio/sha256-simd/cpuid_amd64.go
… vendor/github.com/minio/sha256-simd/cpuid_amd64.s
… vendor/github.com/minio/sha256-simd/cpuid_arm.go
… vendor/github.com/minio/sha256-simd/cpuid_linux_arm64.go
… vendor/github.com/minio/sha256-simd/cpuid_other.go
… vendor/github.com/minio/sha256-simd/go.mod
… vendor/github.com/minio/sha256-simd/sha256.go
… vendor/github.com/minio/sha256-simd/sha256blockAvx2_amd64.go
… vendor/github.com/minio/sha256-simd/sha256blockAvx2_amd64.s
… vendor/github.com/minio/sha256-simd/sha256blockAvx512_amd64.asm
… vendor/github.com/minio/sha256-simd/sha256blockAvx512_amd64.go
… vendor/github.com/minio/sha256-simd/sha256blockAvx512_amd64.s
… vendor/github.com/minio/sha256-simd/sha256blockAvx_amd64.go
… vendor/github.com/minio/sha256-simd/sha256blockAvx_amd64.s
… vendor/github.com/minio/sha256-simd/sha256blockSha_amd64.go
… vendor/github.com/minio/sha256-simd/sha256blockSha_amd64.s
… vendor/github.com/minio/sha256-simd/sha256blockSsse_amd64.go
… vendor/github.com/minio/sha256-simd/sha256blockSsse_amd64.s
… vendor/github.com/minio/sha256-simd/sha256block_amd64.go
… vendor/github.com/minio/sha256-simd/sha256block_arm64.go
… vendor/github.com/minio/sha256-simd/sha256block_arm64.s
… vendor/github.com/minio/sha256-simd/sha256block_other.go
… vendor/github.com/minio/sha256-simd/test-architectures.sh
… vendor/vendor.json