I did run a cpu-profile run on our master implementation, and found this:

```
File: mender-artifact
Build ID: c687cacd156a1724313d2b22b2377fc73952fdbc
Type: cpu
Time: Apr 16, 2020 at 11:11am (CEST)
Duration: 9.54mins, Total samples = 5.77mins (60.49%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top10 
Showing nodes accounting for 286.34s, 82.66% of 346.39s total
Dropped 129 nodes (cum <= 34.64s)
Showing top 10 nodes out of 34
      flat  flat%   sum%        cum   cum%
   132.65s 38.29% 38.29%    251.90s 72.72%  compress/flate.(*compressor).deflate
    53.72s 15.51% 53.80%     53.72s 15.51%  crypto/sha256.block
    47.52s 13.72% 67.52%     47.63s 13.75%  compress/flate.(*compressor).findMatch
    42.10s 12.15% 79.68%     42.97s 12.41%  syscall.Syscall
     9.61s  2.77% 82.45%     56.51s 16.31%  compress/flate.(*huffmanBitWriter).indexTokens
     0.33s 0.095% 82.55%     47.35s 13.67%  compress/flate.(*huffmanEncoder).generate
     0.13s 0.038% 82.58%     67.48s 19.48%  compress/flate.(*huffmanBitWriter).writeBlock
     0.13s 0.038% 82.62%    342.06s 98.75%  io.copyBuffer
     0.08s 0.023% 82.64%    268.32s 77.46%  archive/tar.(*regFileWriter).Write
     0.07s  0.02% 82.66%    268.39s 77.48%  archive/tar.(*Writer).Write
```

Where you see that the most time is spent in compression, and the second most time is spent in shasumming. 

The reason that this targets the second most costly functionality is simply that it was such an easy fix to add.

[here](https://github.com/minio/sha256-simd/) is a link to the library added.

And after the addition:

```
File: mender-artifact
Build ID: 7a3c432c10ea12425e0cc0697d613012dc07d9d0
Type: cpu
Time: Apr 16, 2020 at 11:30am (CEST)
Duration: 8.60mins, Total samples = 5.57mins (64.80%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top10
Showing nodes accounting for 277.71s, 83.05% of 334.39s total
Dropped 146 nodes (cum <= 33.44s)
Showing top 10 nodes out of 31
      flat  flat%   sum%        cum   cum%
   134.22s 40.14% 40.14%    249.19s 74.52%  compress/flate.(*compressor).deflate
    47.85s 14.31% 54.45%     47.85s 14.31%  github.com/mendersoftware/mender-artifact/vendor/github.com/minio/sha256-simd.blockAvx2
    40.13s 12.00% 66.45%     40.18s 12.02%  compress/flate.(*compressor).findMatch
    38.92s 11.64% 78.09%     39.78s 11.90%  syscall.Syscall
    15.90s  4.75% 82.84%     59.86s 17.90%  compress/flate.(*huffmanBitWriter).indexTokens
     0.35s   0.1% 82.95%     44.36s 13.27%  compress/flate.(*huffmanEncoder).generate
     0.11s 0.033% 82.98%    283.91s 84.90%  io.copyBuffer
     0.09s 0.027% 83.01%    264.28s 79.03%  archive/tar.(*Writer).Write
     0.08s 0.024% 83.03%    264.19s 79.01%  archive/tar.(*regFileWriter).Write
     0.06s 0.018% 83.05%     70.24s 21.01%  compress/flate.(*huffmanBitWriter).writeBlock
```